fix tests after architecture change
