+++ lib/resolvers/checkers/index.ts
@@ -1,0 +1,2 @@
+export { default as checkEmail } from "./email";
+export { default as checkPassword } from "./password";
